requirements.txt
astroid==2.4.1
atomicwrites==1.4.0
attrs==19.3.0
bandit==1.6.2
certifi==2019.9.11
chardet==3.0.4
click==7.1.2
colorama==0.4.3
Django==2.2.6
django-debug-toolbar==3.1.1
djangorestframework==3.12.2
gitdb==4.0.5
GitPython==3.1.3
idna==2.8
importlib-metadata==1.5.0
isort==4.3.21
lazy-object-proxy==1.4.3
mccabe==0.6.1
more-itertools==8.2.0
mypy==0.780
mypy-extensions==0.4.3
packaging==20.1
pbr==5.4.5
Pillow==7.0.0
pip-tools==5.3.1
pluggy==0.13.1
py==1.8.1
pycodestyle==2.6.0
pydocstyle==5.0.2
pyflakes==2.2.0
pylama==7.7.1
pylint==2.5.2
pylint-django==2.0.15
pylint-plugin-utils==0.6
pyparsing==2.4.6
pytest==5.3.5
pytest-django==3.8.0
python-dateutil==2.8.1
python-environ==0.4.54
pytz==2019.3
PyYAML==5.3.1
requests==2.22.0
six==1.14.0
smmap==3.0.4
snowballstemmer==2.0.0
sorl-thumbnail==12.6.3
sqlparse==0.3.0
stevedore==2.0.0
toml==0.10.0
typed-ast==1.4.1
typing-extensions==3.7.4.2
urllib3==1.25.6
wcwidth==0.1.8
wrapt==1.12.1
zipp==2.2.0
pytils==0.3
python-dotenv==0.15.0
gunicorn